requirements.txt
aiohttp==3.7.4.post0
aiohttp-retry==2.4.5
appdirs==1.4.4
argon2-cffi==21.1.0
async-timeout==3.0.1
atomicwrites==1.4.0
atpublic==2.3
attrs==21.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bleach==4.1.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
cryptography==3.4.8
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.2.1
distlib==0.3.2
distro==1.6.0
dpath==2.0.5
dulwich==0.20.25
dvc==2.7.4
entrypoints==0.3
filelock==3.0.12
flake8==3.9.2
Flask==2.0.1
Flask-Cors==3.0.10
flatten-dict==0.4.2
flufl.lock==3.2
fsspec==2021.8.1
ftfy==6.0.3
funcy==1.16
future==0.18.2
gitdb==4.0.7
GitPython==3.1.24
google-api-core==2.0.1
google-api-python-client==2.21.0
google-auth==2.1.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
grandalf==0.6
gunicorn==20.1.0
httplib2==0.19.1
idna==3.2
iniconfig==1.1.1
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.3
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
mailchecker==4.0.11
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
multidict==5.1.0
nanotime==0.5.2
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.4
numpy==1.21.2
oauth2client==4.1.3
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.8.1
phonenumbers==8.12.31
pickleshare==0.7.5
platformdirs==2.3.0
pluggy==1.0.0
ply==3.11
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.0
psutil==5.8.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydot==1.4.2
PyDrive2==1.9.4
pyflakes==2.3.1
pygit2==1.6.1
Pygments==2.10.0
pygtrie==2.4.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
python-benedict==0.24.2
python-dateutil==2.8.2
python-fsutil==0.5.0
python-slugify==5.0.2
pytz==2021.1
pywin32==225
pywinpty==1.1.4
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.1
requests==2.26.0
rich==10.10.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
scikit-learn==0.24.2
scipy==1.7.1
Send2Trash==1.8.0
shortuuid==1.0.1
shtab==1.4.1
six==1.16.0
sklearn==0.0
smmap==4.0.0
-e git+https://github.com/jayaram87/dvc_basics.git@6d7a249e16e34e2031209699b81814f0373cea73#egg=src
tabulate==0.8.9
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==2.2.0
toml==0.10.2
tornado==6.1
tox==3.24.4
tqdm==4.62.2
traitlets==5.1.0
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.8.0
voluptuous==0.12.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
win-unicode-console==0.5
wincertstore==0.2
xmltodict==0.12.0
yarl==1.6.3
zc.lockfile==2.0
